+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/models/MetadataReader.kt
@@ -1,14 +1,29 @@
-package com.cyanchill.missingcore.metadataretriever.utils
+package com.cyanchill.missingcore.metadataretriever.models
 
 import android.media.MediaMetadataRetriever
+import android.util.Base64
 import androidx.annotation.OptIn
 import androidx.media3.common.Format
+import androidx.media3.common.MimeTypes
+import androidx.media3.common.PercentageRating
+import androidx.media3.common.Rating
 import androidx.media3.common.MediaMetadata as AndroidXMediaMetadata
 import androidx.media3.common.util.UnstableApi
+import java.net.URLConnection
 
+import com.cyanchill.missingcore.metadataretriever.utils.NormalizationUtils
+
+/**
+ * Utilities to format & normalize sources of metadata as a map.
+ */
 @OptIn(UnstableApi::class)
-object MetadataMapperUtils {
-  fun mapFormatMetadata(format: Format): HashMap<String, Any?> {
+object MetadataReader {
+  /**
+   * Relevant metadata fields found on `Format`.
+   *
+   * @see <a href="https://developer.android.com/reference/androidx/media3/common/Format">Link</a>
+   */
+  fun fromFormat(format: Format): HashMap<String, Any?> {
     val dataMap = hashMapOf<String, Any?>()
 
     // Set `Int?` values.
@@ -22,14 +37,19 @@ object MetadataMapperUtils {
     return dataMap
   }
 
-  fun mapMediaMetadata(
+  /**
+   * All metadata fields found on `MediaMetadata`.
+   *
+   * @see <a href="https://developer.android.com/reference/androidx/media3/common/MediaMetadata">Link</a>
+   */
+  fun fromMediaMetadata(
     mediaMetadata: AndroidXMediaMetadata,
     getArtworkData: Boolean = false,
   ): HashMap<String, Any?> {
     val dataMap = hashMapOf<String, Any?>()
 
     // Pre-compute values to put in hash map.
-    val artworkData = if (getArtworkData) MediaMetadataUtils.getBase64Image(mediaMetadata.artworkData) else null
+    val artworkData = if (getArtworkData) getBase64Image(mediaMetadata.artworkData) else null
     val trackNumber = if (mediaMetadata.trackNumber == 0) null else mediaMetadata.trackNumber
     val year = NormalizationUtils.parseYear(mediaMetadata.recordingYear)
       ?: NormalizationUtils.parseYear(mediaMetadata.releaseYear)
@@ -38,8 +58,8 @@ object MetadataMapperUtils {
     dataMap.put("isBrowsable", mediaMetadata.isBrowsable)
     dataMap.put("isPlayable", mediaMetadata.isPlayable)
     // Set `Double?` values.
-    dataMap.put("overallRating", MediaMetadataUtils.getPercentageRating(mediaMetadata.overallRating))
-    dataMap.put("userRating", MediaMetadataUtils.getPercentageRating(mediaMetadata.userRating))
+    dataMap.put("overallRating", getPercentageRating(mediaMetadata.overallRating))
+    dataMap.put("userRating", getPercentageRating(mediaMetadata.userRating))
     // Set `Int?` values.
     dataMap.put("discNumber", mediaMetadata.discNumber)
     dataMap.put("recordingDay", mediaMetadata.recordingDay)
@@ -57,7 +77,7 @@ object MetadataMapperUtils {
     dataMap.put("albumTitle", mediaMetadata.albumTitle?.toString())
     dataMap.put("artist", mediaMetadata.artist?.toString())
     dataMap.put("artworkData", artworkData)
-    dataMap.put("artworkDataType", MediaMetadataUtils.getID3PictureType(mediaMetadata.artworkDataType))
+    dataMap.put("artworkDataType", getID3PictureType(mediaMetadata.artworkDataType))
     dataMap.put("artworkUri", mediaMetadata.artworkUri?.toString())
     dataMap.put("compilation", mediaMetadata.compilation?.toString())
     dataMap.put("composer", mediaMetadata.composer?.toString())
@@ -65,7 +85,7 @@ object MetadataMapperUtils {
     dataMap.put("description", mediaMetadata.description?.toString())
     dataMap.put("displayTitle", mediaMetadata.displayTitle?.toString())
     dataMap.put("genre", mediaMetadata.genre?.toString())
-    dataMap.put("mediaType", MediaMetadataUtils.getMediaType(mediaMetadata.mediaType))
+    dataMap.put("mediaType", getMediaType(mediaMetadata.mediaType))
     dataMap.put("station", mediaMetadata.station?.toString())
     dataMap.put("subtitle", mediaMetadata.subtitle?.toString())
     dataMap.put("title", mediaMetadata.title?.toString())
@@ -74,14 +94,19 @@ object MetadataMapperUtils {
     return dataMap
   }
 
-  fun mapMediaMetadataRetriever(
+  /**
+   * Subset of metadata fields found on `MediaMetadataRetriever` that correlates with the fields on `MediaMetadata`.
+   *
+   * @see <a href="https://developer.android.com/reference/androidx/media3/common/MediaMetadata">Link</a>
+   */
+  fun fromMediaMetadataRetriever(
     mmr: MediaMetadataRetriever,
     getArtworkData: Boolean = false,
   ): HashMap<String, Any?> {
     val dataMap = hashMapOf<String, Any?>()
 
     // Pre-compute values to put in hash map.
-    val artworkData = if (getArtworkData) MediaMetadataUtils.getBase64Image(mmr.getEmbeddedPicture()) else null
+    val artworkData = if (getArtworkData) getBase64Image(mmr.getEmbeddedPicture()) else null
     val trackNumber = mmr.extractMetadata(MediaMetadataRetriever.METADATA_KEY_CD_TRACK_NUMBER)
       ?.let { if (it.toIntOrNull() == 0) null else it.toIntOrNull() }
     val year = NormalizationUtils.parseYear(mmr.extractMetadata(MediaMetadataRetriever.METADATA_KEY_YEAR)) ?: run {
@@ -114,4 +139,109 @@ object MetadataMapperUtils {
 
     return dataMap
   }
+
+  //#region [Exposed Helpers To Parse Metadata Values]
+  /** Returns a base64 image string from a `ByteArray`. */
+  fun getBase64Image(bytes: ByteArray? = null, maxSizeMB: Double = 5.0): String? {
+    if (bytes == null) return null
+    // Determine the mimetype from bytes.
+    val mimeType = URLConnection.guessContentTypeFromStream(bytes.inputStream())?.let {
+      MimeTypes.normalizeMimeType(it)
+    }
+    // Ensure the mimeType we get is defined and is for an image.
+    if (!MimeTypes.isImage(mimeType)) return null
+    // Convert max MB to bytes. We take 3/4 of the max MB as converting a byte array to a base64
+    // string causes a 33% increase in size.
+    val maxSizeBytes = maxSizeMB * 0.75 * 1024 * 1024
+    if (bytes.size > maxSizeBytes) return null
+    return "data:$mimeType;base64,${Base64.encodeToString(bytes, Base64.DEFAULT)}"
+  }
+  //#endregion
+
+  //#region [Internal Helpers To Parse Metadata Values]
+  /**
+   * Convert integer picture type to a human-readable string.
+   *
+   * @see <a href="https://developer.android.com/reference/androidx/media3/common/MediaMetadata.PictureType">Link</a>
+   */
+  private fun getID3PictureType(code: Int?): String? = when (code) {
+    0 -> "Other"
+    1 -> "32x32 pixels 'file icon' (PNG only)"
+    2 -> "Other file icon"
+    3 -> "Cover (front)"
+    4 -> "Cover (back)"
+    5 -> "Leaflet page"
+    6 -> "Media (e.g. label side of CD)"
+    7 -> "Lead artist/lead performer/soloist"
+    8 -> "Artist/performer"
+    9 -> "Conductor"
+    10 -> "Band/Orchestra"
+    11 -> "Composer"
+    12 -> "Lyricist/text writer"
+    13 -> "Recording Location"
+    14 -> "During recording"
+    15 -> "During performance"
+    16 -> "Movie/video screen capture"
+    17 -> "A bright coloured fish"
+    18 -> "Illustration"
+    19 -> "Band/artist logotype"
+    20 -> "Publisher/Studio logotype"
+    else -> null
+  }
+
+  /**
+   * Convert integer media type to a human-readable string.
+   *
+   * @see <a href="https://developer.android.com/reference/androidx/media3/common/MediaMetadata.MediaType">Link</a>
+   */
+  private fun getMediaType(code: Int?): String? = when (code) {
+    0 -> "Mixed"
+    1 -> "Music"
+    2 -> "Audio book chapter"
+    3 -> "Podcast episode"
+    4 -> "Radio station"
+    5 -> "News"
+    6 -> "Video"
+    7 -> "Trailer"
+    8 -> "Movie"
+    9 -> "TV show"
+    10 -> "Album"
+    11 -> "Artist"
+    12 -> "Genre"
+    13 -> "Playlist"
+    14 -> "Year"
+    15 -> "Audio book"
+    16 -> "Podcast"
+    17 -> "TV channel"
+    18 -> "TV series"
+    19 -> "TV season"
+    20 -> "Folder mixed"
+    21 -> "Folder albums"
+    22 -> "Folder artists"
+    23 -> "Folder genres"
+    24 -> "Folder playlists"
+    25 -> "Folder years"
+    26 -> "Folder audio books"
+    27 -> "Folder podcasts"
+    28 -> "Folder tv channels"
+    29 -> "Folder tv series"
+    30 -> "Folder tv shows"
+    31 -> "Folder radio stations"
+    32 -> "Folder news"
+    33 -> "Folder videos"
+    34 -> "Folder trailers"
+    35 -> "Folder movies"
+    else -> null
+  }
+
+  /**
+   * Get the percentage rating from a `Rating`.
+   *
+   * @see <a href="https://developer.android.com/reference/androidx/media3/common/Rating">Link</a>
+   */
+  private fun getPercentageRating(rating: Rating?): Double? = when (rating?.isRated()) {
+    true -> PercentageRating.fromBundle(rating.toBundle()).getPercent().toDouble()
+    else -> null
+  }
+  //#endregion
 }
